Create an eth address per BNB address
So on create swap, we check to see if the BNB address for that token exists.
If so, we return the eth address
If not, we create a new address then return it.

On finalize.
Take the token with the eth address, check to see if a transfer was done.
Validate that against the swaps that have been recorded previously.
Insert all new deposits into swaps.
Return all new deposits.

Signed-off-by: anton <anton@mfin.co.za>